Move `_handle_keyboard_interrupt` method
